[GH-6] Transition script & tweak params.
I added a transition script. It will take the old meta data format and
return the new one. To get the new one, the script will guess at how
many times "yes" was answered and how many times "no" was answered. It's
impossible to get those numbers exact because of information loss, but
it is calculated to be believable values. The current date is also
calculated. That is an exact calculation. Also dates are all switched to
UTC RFC3339 format. And the rank is removed.

I also made a parameter tweak to the sm2 algorithm. I added or
subtracted .01 from all the weights when calculating the rank, so that
the rank isn't always just a multiple of .5.