GH-46 GH-47 GH-48 Refactor up changes.
The program at least compiles completely now after I refactored the
previous changes throughout the project. As part of this, I removed the
"forget" and "skip" functionality from the UI. I found that I never used
those. But also, the new way cards are passed off will not depend on
deck order at all, but timestamps instead. This makes the "skip"
functionality useless. And the forget functionality is just ugly. You
should just remove the card or not remove the file. Not having the
"forget" command encourages you to keep better care of your notes.