Got the new deck api compiling.
Yesterday I made a list of things I need to do to call this refactor
complete. This list is in #48.

1. Make sure the current tests are working again.
2. Review the deck API again.
3. Refactor the logic for the filters & maybe remove the drop function.
4. Start writing/resurrecting tests for the deck.
5. Once I feel like the deck is solid, review the file package.
6. Make sure I have near 100% coverage in the card, meta, & plugin packages again.
7. Go over any termbox UI changes I need to make. As well as main.go changes.

I pretty much did 1, 2, 3, and 5 today. Time to make a new list though!

1. Test the deck package.
2. Test the file package.
3. Double check all the test coverage.
4. Tiny termbox UI change.
5. Manual testing to make sure everything still works.

I won't actually add the "outcome" file & logic on this branch before
the merge. I'll do that once these changes are stable and on master.