Working on add predict tests.
AddPredicts() will add a predict if it doesn't already exist, or it
exists and the value is zero. This commit fixed a few bugs around the
predict adding logic.

I'm thinking I will want to improve it more by checking the Total, and
if the total is less than or equal to the new predict's total, then
overwrite the predict.