File Tests. Project builds.
I finally redid the file tests. Files were taking in a deck before, but
now they don't and they return lists. This is a much better approach,
because the deck object has a lot of extra information.

This commit builds, but doesn't run. There is a problem with the deck
filter logic. I need to add tests for that next and fix the problem.