Fixed the exec bug & added unit test.
Exec wasn't adding the cards from the future deck to the review deck
correctly. I fixed that and added unit tests to the stack & deck
modules. It seems like editting is working, so this probably fixed the
editting bug too, but I should still make a unit test for Edit to make
sure.