[GH-51] Fixed streak.
Streak will now always be something that could be valid. This introduced
a slight change in functionality too. Before, specifying the streak
could also change the yes/no count. Now the yes and no count instead
could change the streak. This is both because I don't want to trouble
with yes/no counts going above the max, but also because my
understanding of the concept of yes and no is better. The yes/no counts
are fundamentally more important than the streak count. They are what's
used to calculate the total. And they are more reliable.