[GH-66] Spaces don't matter for parser.
The escaping strategy is better now too, you can escape any character.
You can also escape the starting `@>` token. What I'm most proud of
about this is the speed. My first solution was slow, because I was doing
string and rune casting in order to keep track of things. This solution
is just as fast, or faster, than the parse words, because it is using
the same SplitFunc interface. It might be slower than parsewords since
there is technically more logic, but it might be faster, because it
parses larger areas of text at a time.