[GH-71] Fixed cloze problems.
While writing the README, I wrote some examples that were actually
broken in concards. I added those cases as tests and fixed the problems.
Now that cloze cards are working. That readme was kind of like writing a
big test. Now that cloze is working, more time could be spent
refactoring it.